chore(proxy) Dry http project registration handler
